+++ src/main/java/card/CardNumber.java
@@ -1,4 +1,4 @@
-package model.card;
+package card;
 
 import java.util.Collections;
 import java.util.List;
