+++ src/main/java/card/CardType.java
@@ -1,4 +1,4 @@
-package model.card;
+package card;
 
 public enum CardType {
