--- src/main/java/constant/BlackjackConstant.java
@@ -1,10 +1,0 @@
-package constant;
-
-public class BlackjackConstant {
-
-    public static final int BUST_NUMBER = 21;
-    public static final int DEALER_CARDS_MIN_TOTAL = 17;
-
-    private BlackjackConstant() {
-    }
-}
