+++ src/main/java/deck/ShuffledDeckCreator.java
@@ -1,11 +1,11 @@
-package model.deck;
+package deck;
 
+import card.Card;
+import card.CardNumber;
+import card.CardType;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import model.card.Card;
-import model.card.CardNumber;
-import model.card.CardType;
 
 public class ShuffledDeckCreator implements DeckCreateStrategy {
