+++ src/test/java/domain/deck/DeckTest.java
@@ -1,5 +1,6 @@
-package model.deck;
+package domain.deck;
 
+import deck.Deck;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
